requirements.txt
anyio==3.6.2
APScheduler==3.10.1
asgiref==3.6.0
certifi==2022.12.7
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
instagram-private-api==1.6.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2023.3
tzlocal==4.3
